Integrate JSON Schema Form editor
Integrates the JSON Schema form builder available at https://github.com/backoffice-plus/formbuilder-embedded/
into the plugin. It ships a new WordPress plugin with a build pipeline
and a pre-built version of the form builder.

 ## How does it work?

This PR adds an `iframe` that embeds a standalone HTML page with the
form builder component in it. It then uses postMessage and onmessage to
send any JSON Schema updates to the form and retrieve them from the form.

 ## Follow-up work

* Connect the vite build pipeline into the main pnpm build flow.
* Remove the pre-built artifacts from the dist/ directory.
* Make the formbuilder-embedded packages importable in the rest of the
  code in this repo.
* Perhaps? Remove the iframe and embed the components directly.